test(feed-service,subscription-service): migrate service-layer tests to BDD and integration coverage - add Cucumber test runners, feature files, step definitions, and object factories for feed-service and subscription-service - introduce shared BDD and integration test constants/factories to reduce duplication and stabilize test fixtures - add integration test suites for both services focused on real persistence and service interaction paths - remove obsolete unit tests and old test utility classes that are superseded by the new BDD/integration approach - update test dependencies/configuration in both build scripts to support Cucumber and integration test execution

+++ feed_service/src/main/java/com/kolmir/feed_service/controller/CommentController.java
@@ -38,7 +38,7 @@ public class CommentController {
     }
 
     @GetMapping(COMMENTS_COUNT_FOR_POST)
-    public ResponseEntity<Integer> getCountForPost(@PathVariable Long postId) {
+    public ResponseEntity<Long> getCountForPost(@PathVariable Long postId) {
         return ResponseEntity.ok(commentService.getCommentsCountForPost(postId));
     }
