+++ feed_service/src/main/java/com/kolmir/feed_service/service/CommentService.java
@@ -9,7 +9,7 @@ import com.kolmir.feed_service.dto.comment.*;
 
 public interface CommentService {
     public Page<CommentResponse> getAllCommentsForPost(Long postId, Pageable pageable);
-    public Integer getCommentsCountForPost(Long postId);
+    public Long getCommentsCountForPost(Long postId);
     public Page<CommentResponse> getAllFromUserToPost(Long userId, Long postId, Pageable pageable);
     public CommentResponse getById(Long commentId);
     public CommentResponse create(CommentCreateRequest request);
